Create getStrokeHitbox util
Create getStrokeHitbox util to extract hitbox related logic in one spot
